pyproject.toml
[project]
name = "file-client"
version = "0.1"
description = "CLI application which retrieves and prints data from one of the described backends"
readme = "README.md"
authors = [{name = "Lukáš Říha"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "requests",
    "grpcio",
    "grpcio-tools",
]

[project.urls]
Homepage = "https://github.com/cedel1/file-client"
Changelog = "https://github.com/cedel1/file-client/releases"
Issues = "https://github.com/cedel1/file-client/issues"
CI = "https://github.com/cedel1/file-client/actions"

[project.entry-points.console_scripts]
file-client = "file_client.cli:cli"

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "responses"]
